+++ hdrop-web-next/src/app/layout.tsx
@@ -1,7 +1,6 @@
 import { Inter } from 'next/font/google'
 
-import Header from '@/components/Header'
-import Footer from '@/components/Footer'
+import { Header, Footer } from '@/components'
 
 import './globals.css'
